refactor: refactor dashboard
