+++ app/api/projects/route.ts
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth/next";
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
-import { authOptions } from "@/app/api/auth/[...nextauth]/auth";
+import { authOptions } from "@/lib/auth.config";
 
 export async function POST(request: Request) {
 	const session = await getServerSession(authOptions);
