+++ app/dashboard/page.tsx
@@ -1,0 +1,8 @@
+import Dashboard from "@/components/dashboard";
+import React from "react";
+
+const page = () => {
+	return <Dashboard />;
+};
+
+export default page;
