+++ app/projects/new/page.tsx
@@ -1,7 +1,7 @@
 import { ProjectForm } from "@/components/project-form";
 import React from "react";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/auth";
+import { authOptions } from "@/lib/auth.config";
 
 const Page = async () => {
 	const session = await getServerSession(authOptions);
